deviantart::types::deviation

Struct Deviation

pub struct Deviation {
    pub block_reasons: Vec<Value>,
    pub deviation_id: u64,
    pub kind: String,
    pub url: Url,
    pub media: Media,
    pub title: String,
    pub text_content: Option<TextContext>,
    pub is_downloadable: bool,
    pub unknown: HashMap<String, Value>,
}

A Deviation

Fields
block_reasons: Vec<Value>

?

deviation_id: u64

Deviation ID

kind: String

Deviation Type

url: Url

Image Url

media: Media

Media Info

title: String

Title

text_content: Option<TextContext>

Text content for literature

is_downloadable: bool

Whether this is downloadable

unknown: HashMap<String, Value>

Unknown K/Vs

Implementations

impl Deviation

pub fn get_media_url(&self) -> Option<Url>

Get the media url for this Deviation.

pub fn get_download_url(&self) -> Option<Url>

Get the “download” url for this Deviation.

pub fn get_fullview_url( &self, options: GetFullviewUrlOptions, ) -> Result<Url, GetFullviewUrlError>

Get the fullview url for this Deviation.

pub fn get_gif_url(&self) -> Option<Url>

Get the GIF url for this Deviation.

pub fn get_best_video_url(&self) -> Option<&Url>

Get the best video url

pub fn is_image(&self) -> bool

Whether this is an image

pub fn is_literature(&self) -> bool

Whether this is literature

pub fn is_film(&self) -> bool

Whether this is a film

pub fn get_image_download_url(&self) -> Option<Url>

Get the most “fitting” url to download an image.

Usually, [DeviationExtended] holds better data than a Deviation, so prefer that instead.

pub fn get_extension(&self) -> Option<&str>

Try to get the original extension of this Deviation